Corrected long-standing bug in the LinkDoors algorithm
Bug was present on the original id code. Not every door touching would be added to the same group, but this flaw wasn't apparent due to the fact that all doors moved at the same time. When delayed opening/closing was introduced, this became a problem.
So the algorithm was completely redone.
- LinkDoors loop is now fired as a single function from the worldspawn, instead of from each door.
- A processing queue was created to prevent isolated linked door pairs from creating a new group.