Rewritten weapons/items flag system
Removed extra_items and added a weapons flag. I moved all weapons into that new flag, so I could remove that inconsistency between IT_ and WEAPON_ for the new weapons.

Everything is just WEAPON_* now, and doubles as a slot in the weapons flag and in weapon number, just like the original weapon values.

The only small catch is that the OG weapons shown in the HUD are hardwired into  items and the original flag values by the engine. so the 1-64 slots in items are reserved, and I change that to match the weapons flag for that range every time a weapon is added/removed, which is basically on level load, on weapon pickup, and trigger_remove_items.