Added lock for doors through target_setstate
Similar to Copper's method.
Locking a door will prevent it from opening in any way.
If the door's open it'll close immediately.
If the door is togglable or has wait = -1 (including key doors), it'll remain in its position when disabled. You can alternatively force it to close anyway with a spawnflag, and it'll open again when unlocked.